% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertAsTibbleAddin}
\alias{insertAsTibbleAddin}
\title{Insert as_tibble}
\usage{
insertAsTibbleAddin()
}
\description{
Call this function as an addin to insert \code{as_tibble(} at the cursor position.
}

