% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertDplyrAddin}
\alias{insertDplyrAddin}
\title{Insert dplyr::}
\usage{
insertDplyrAddin()
}
\description{
Call this function as an addin to insert \code{dplyr::} at the cursor position.
}

