% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertFilterAddin}
\alias{insertFilterAddin}
\title{Insert filter.}
\usage{
insertFilterAddin()
}
\description{
Call this function as an addin to insert \code{filter(} at the cursor position.
}

