% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertGatherAddin}
\alias{insertGatherAddin}
\title{Insert gather.}
\usage{
insertGatherAddin()
}
\description{
Call this function as an addin to insert \code{gather(} at the cursor position.
}

