% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertGroupByAddin}
\alias{insertGroupByAddin}
\title{Insert group_by}
\usage{
insertGroupByAddin()
}
\description{
Call this function as an addin to insert \code{group_by(} at the cursor position.
}

