% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertLeftJoinAddin}
\alias{insertLeftJoinAddin}
\title{Insert left_join.}
\usage{
insertLeftJoinAddin()
}
\description{
Call this function as an addin to insert \code{left_join(} at the cursor position.
}

