% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertLubridateAddin}
\alias{insertLubridateAddin}
\title{Insert lubridate::}
\usage{
insertLubridateAddin()
}
\description{
Call this function as an addin to insert \code{lubridate::(} at the cursor position.
}

