% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertMutateAddin}
\alias{insertMutateAddin}
\title{Insert mutate.}
\usage{
insertMutateAddin()
}
\description{
Call this function as an addin to insert \code{mutate(} at the cursor position.
}

