% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertPurrrAddin}
\alias{insertPurrrAddin}
\title{Insert purrr::.}
\usage{
insertPurrrAddin()
}
\description{
Call this function as an addin to insert \code{purrr::} at the cursor position.
}

