% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertSelectAddin}
\alias{insertSelectAddin}
\title{Insert select}
\usage{
insertSelectAddin()
}
\description{
Call this function as an addin to insert \code{select(} at the cursor position.
}

