% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertSpreadAddin}
\alias{insertSpreadAddin}
\title{Insert spread.}
\usage{
insertSpreadAddin()
}
\description{
Call this function as an addin to insert \code{spread(} at the cursor position.
}

