% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertStringrAddin}
\alias{insertStringrAddin}
\title{Insert stringr::}
\usage{
insertStringrAddin()
}
\description{
Call this function as an addin to insert \code{stringr::(} at the cursor position.
}

