% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertSummarizeAddin}
\alias{insertSummarizeAddin}
\title{Insert summarize()}
\usage{
insertSummarizeAddin()
}
\description{
Call this function as an addin to insert \code{summarize(} at the cursor position.
}

