% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertTidyrAddin}
\alias{insertTidyrAddin}
\title{Insert tidyr::.}
\usage{
insertTidyrAddin()
}
\description{
Call this function as an addin to insert \code{tidyr::} at the cursor position.
}

