% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addin_defs.R
\name{insertUnnestAddin}
\alias{insertUnnestAddin}
\title{Insert unnest.}
\usage{
insertUnnestAddin()
}
\description{
Call this function as an addin to insert \code{unnest(} at the cursor position.
}

